refactor(config): split writer by namespace, extract SettingsSection
Address low-severity review nits on the #12 branch:
- Split generalBehavior() into mouseResize() and animations() so
  each writer function maps to one Lua namespace, matching the
  per-concern structure of the other sections.
- Move the shared SettingsSection view into its own file so
  PresetsTab no longer implicitly depends on GeneralTab.
